+++ src/main/java/shuheng/Parser.java
@@ -1,14 +1,14 @@
-package duke;
+package shuheng;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
 
-import duke.exceptions.HistoryIndexException;
-import duke.exceptions.InvalidDateTimeException;
-import duke.exceptions.InvalidInputException;
-import duke.exceptions.InvalidParametersException;
-import duke.exceptions.InvalidTaskException;
-import duke.tasks.PriorityLevel;
+import shuheng.exceptions.HistoryIndexException;
+import shuheng.exceptions.InvalidDateTimeException;
+import shuheng.exceptions.InvalidInputException;
+import shuheng.exceptions.InvalidParametersException;
+import shuheng.exceptions.InvalidTaskException;
+import shuheng.tasks.PriorityLevel;
 
 /**
  * This class represents the parser used to receive and handle user inputs to the UI.
