+++ src/main/java/shuheng/ShuHengLauncher.java
@@ -1,11 +1,11 @@
-package duke;
+package shuheng;
 
 import javafx.application.Application;
 
 /**
  * This class represents the launcher for duke!
  */
-public class DukeLauncher {
+public class ShuHengLauncher {
     public static void main(String[] args) {
         Application.launch(Main.class, args);
     }
