+++ src/main/java/shuheng/exceptions/CorruptedLogException.java
@@ -1,4 +1,4 @@
-package duke.exceptions;
+package shuheng.exceptions;
 
 /**
  * This class represents the error thrown with a corrupted log.
