+++ src/main/java/shuheng/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package duke.exceptions;
+package shuheng.exceptions;
 
 /**
  * This class represents exceptions thrown by Duke, our friendly chatbot.
