+++ src/main/java/shuheng/exceptions/InvalidDateTimeException.java
@@ -1,4 +1,4 @@
-package duke.exceptions;
+package shuheng.exceptions;
 
 /**
  * This class represents the error caused via an invalid date time.
