+++ src/main/java/shuheng/frontend/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.frontend;
+package shuheng.frontend;
 
 import java.io.IOException;
 import java.util.Collections;
