+++ src/main/java/shuheng/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package shuheng.tasks;
 
 import java.time.LocalDateTime;
