+++ src/main/java/shuheng/tasks/Event.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package shuheng.tasks;
 
 import java.time.LocalDateTime;
