+++ src/main/java/shuheng/tasks/PriorityLevel.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package shuheng.tasks;
 
 /**
  * This class is a enum representing priority levels for tasks given to duke.
